Update Coffee2.html
I have implemented a few additions of SEO features into my website. I have also added "lazy" loading as an added optimisation feature. There is a drawback to this for SEO, which is that the browser could load the website too fast and this would result in waiting for images to still load in. I have also taken the time to reduce the amount of white space in all my code. In this instance I have not reduced the hex values but for future projects I will definitely reduce the hex values where applicable. I have also added in "lazy".