Create SINISTER - Group Project (Python)
This was the code for the gaming project which I created as a team. I organised the Trello (where we planned the game), I created the narrative driven storyline (the storyline currently used in the game), I helped others with coding issues, I managed the team and I also created the powerpoint slides for the final presentation, effectively making sure we reached the deadline early.